API/requirements.txt
numpy==1.19.4
opencv-python-headless
gunicorn==20.0.4
Pillow==8.1.0
fastapi==0.62.0